pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tk-listing-workflow"
version = "0.1.0"
description = "MVP scaffold for TikTok automated listing workflow"
requires-python = ">=3.11"
dependencies = ["Pillow>=11.0.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
